# storage/api-docs/api-docs.json
{"swagger":"2.0","info":{"title":"L5 Swagger API","description":"L5 Swagger API description","contact":{"email":"konstantinrjab@gmail.com"},"version":"1.0.0"},"host":"127.0.0.1:8000\/api","basePath":"\/","schemes":["http","https"],"paths":{"\/login":{"post":{"tags":["user"],"summary":"Login User","parameters":[{"name":"email","in":"formData","description":"User Email","required":true,"schema":{"example":"admin@test.com"},"type":"string","format":"email"},{"name":"password","in":"formData","description":"User Password","required":true,"schema":{"example":"secret_pass"},"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"properties":{"user":{"type":"object","$ref":"#definitions\/user"}}}},"default":{"description":"Error","schema":{"properties":{"error":{"type":"object","$ref":"#definitions\/error"}}}}}}},"\/categories":{"get":{"tags":["category"],"summary":"Get list of categories","description":"Returns list of categories","responses":{"200":{"description":"successful operation","schema":{"properties":{"categories":{"type":"array","items":{"$ref":"#definitions\/category"}}}}},"400":{"description":"Bad request"}}},"post":{"tags":["category"],"summary":"Create new category","parameters":[{"name":"name","in":"formData","required":true,"schema":{"example":"test name 1"},"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"properties":{"category":{"type":"object","$ref":"#definitions\/category"}}}},"default":{"description":"Error","schema":{"properties":{"error":{"type":"object","$ref":"#definitions\/error"}}}}},"security":[{"api_key":[]}]}},"\/categories\/{categoryID}":{"get":{"tags":["category"],"summary":"Get category","parameters":[{"name":"categoryID","in":"path","required":true,"schema":{"example":"1"},"type":"integer"},{"name":"items","in":"query","schema":{"example":"true"},"type":"boolean"}],"responses":{"200":{"description":"successful operation","schema":{"properties":{"category":{"type":"object","$ref":"#definitions\/category"}}}},"default":{"description":"Error","schema":{"properties":{"error":{"type":"object","$ref":"#definitions\/error"}}}}}},"post":{"tags":["category"],"summary":"Update category","parameters":[{"name":"categoryID","in":"path","required":true,"schema":{"example":"1"},"type":"integer"},{"name":"name","in":"formData","required":true,"schema":{"example":"test name 1"},"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"properties":{"category":{"type":"object","$ref":"#definitions\/category"}}}},"default":{"description":"Error","schema":{"properties":{"error":{"type":"object","$ref":"#definitions\/error"}}}}},"security":[{"api_key":[]}]},"delete":{"tags":["category"],"summary":"Delete category","parameters":[{"name":"categoryID","in":"path","required":true,"schema":{"example":"1"},"type":"integer"}],"responses":{"204":{"description":"successful operation"},"default":{"description":"Error","schema":{"properties":{"error":{"type":"object","$ref":"#definitions\/error"}}}}},"security":[{"api_key":[]}]}},"\/category\/{categoryID}\/parameters":{"get":{"tags":["parameters, category"],"summary":"Get category parameters","description":"Returns list of parameters","operationId":"getCategoryParametersList","responses":{"200":{"description":"successful operation","schema":{"properties":{"count":{"type":"integer"},"parameters":{"type":"array","items":{"$ref":"#definitions\/categoryParameter"}}}}},"400":{"description":"Bad request"}}}},"\/items\/{itemID}\/":{"get":{"tags":["items"],"summary":"Get product","description":"Returns product","operationId":"getItem","parameters":[{"name":"itemID","in":"path","description":"Item ID","required":true,"type":"integer"}],"responses":{"200":{"description":"successful operation","schema":{"properties":{"item":{"type":"object","$ref":"#definitions\/item"}}}},"400":{"description":"Bad request"}}}},"items\/parameters":{"get":{"tags":["parameter"],"summary":"Get list of parameters","description":"Returns list of parameters","responses":{"200":{"description":"successful operation","schema":{"properties":{"count":{"type":"integer"},"parameters":{"type":"array","items":{"$ref":"#definitions\/parameter"}}}}},"400":{"description":"Bad request"}}}},"\/parameters":{"get":{"tags":["parameters"],"summary":"Get list of parameters","description":"Returns list of parameters","operationId":"getParametersList","responses":{"200":{"description":"successful operation","schema":{"properties":{"count":{"type":"integer"},"parameters":{"type":"array","items":{"$ref":"#definitions\/parameter"}}}}},"400":{"description":"Bad request"}}}}},"definitions":{"category":{"properties":{"id":{"description":"Category ID","type":"integer"},"name":{"description":"Category Name","type":"string"},"created_at":{"type":"string"},"updated_at":{"type":"string"},"items_count":{"type":"integer"}}},"error":{"properties":{"message":{"description":"Category ID","type":"string"}}},"image":{"properties":{"id":{"description":"Item ID","type":"integer"},"name":{"description":"Item Name","type":"string"},"category":{"$ref":"#definitions\/category"},"parameters":{"type":"object"},"created_at":{"type":"string"},"updated_at":{"type":"string"}}},"item":{"properties":{"id":{"description":"Item ID","type":"integer"},"name":{"description":"Item Name","type":"string"},"category":{"$ref":"#definitions\/category"},"parameters":{"type":"object"},"created_at":{"type":"string"},"updated_at":{"type":"string"}}},"itemParameter":{"properties":{"id":{"description":"Item ItemParameter ID","type":"integer"},"item_id":{"description":"Item ID","type":"integer"},"parameter_id":{"description":"ItemParameter ID","type":"integer"},"value":{"description":"ItemParameter value","type":"sting"}}},"parameter":{"properties":{"id":{"description":"Parameter ID","type":"integer"},"name":{"description":"Parameter Name","type":"string"},"created_at":{"type":"string"},"updated_at":{"type":"string"}}},"user":{"properties":{"id":{"description":"User ID","type":"integer"},"name":{"description":"User Name","type":"string"},"email":{"description":"User Email","type":"string"},"created_at":{"type":"string"},"updated_at":{"type":"string"},"api_token":{"type":"string"}}}},"tags":[{"name":"category","description":"have products"},{"name":"image","description":"product"},{"name":"items","description":"product"},{"name":"parameters","description":"products have parameters"},{"name":"user"}],"securityDefinitions":{"api_key":{"type":"apiKey","description":"A short description for security scheme","name":"Authorization","in":"header"}}}